move logic (reducer) to separate file
